Tear Time.java apart.
