+++ time/src/main/java/io/spine/time/LocalTimes.java
@@ -20,6 +20,7 @@
 package io.spine.time;
 
 import com.google.protobuf.Timestamp;
+import io.spine.base.Time;
 import io.spine.time.Formats.Parameter;
 
 import java.text.ParseException;
@@ -27,6 +28,8 @@ import java.util.Calendar;
 import java.util.Date;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static io.spine.time.SiTime.MILLIS_PER_SECOND;
+import static io.spine.time.SiTime.NANOS_PER_MILLISECOND;
 import static io.spine.time.Calendars.checkArguments;
 import static io.spine.time.Calendars.getHours;
 import static io.spine.time.Calendars.getMillis;
@@ -34,12 +37,10 @@ import static io.spine.time.Calendars.getMinutes;
 import static io.spine.time.Calendars.getSeconds;
 import static io.spine.time.Calendars.toCalendar;
 import static io.spine.time.Calendars.toLocalTime;
+import static io.spine.time.EarthTime.HOURS_PER_DAY;
+import static io.spine.time.EarthTime.MINUTES_PER_HOUR;
 import static io.spine.time.Formats.appendSubSecond;
 import static io.spine.time.Formats.timeFormat;
-import static io.spine.time.Time.HOURS_PER_DAY;
-import static io.spine.time.Time.MILLIS_PER_SECOND;
-import static io.spine.time.Time.MINUTES_PER_HOUR;
-import static io.spine.time.Time.NANOS_PER_MILLISECOND;
 import static io.spine.validate.Validate.checkBounds;
 import static java.util.Calendar.HOUR;
 import static java.util.Calendar.MILLISECOND;
