+++ time/src/test/java/io/spine/time/Timestamps2Should.java
@@ -32,18 +32,18 @@ import static com.google.protobuf.util.Durations.fromSeconds;
 import static com.google.protobuf.util.Timestamps.add;
 import static com.google.protobuf.util.Timestamps.subtract;
 import static com.google.protobuf.util.Timestamps.toNanos;
+import static io.spine.base.Time.getCurrentTime;
+import static io.spine.base.Time.resetProvider;
+import static io.spine.base.Time.setProvider;
+import static io.spine.base.Time.systemTime;
 import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.time.SiTime.MICROS_PER_SECOND;
+import static io.spine.time.SiTime.MILLIS_PER_SECOND;
+import static io.spine.time.SiTime.NANOS_PER_MICROSECOND;
+import static io.spine.time.SiTime.NANOS_PER_SECOND;
 import static io.spine.time.Durations2.fromMinutes;
-import static io.spine.time.Time.HOURS_PER_DAY;
-import static io.spine.time.Time.MICROS_PER_SECOND;
-import static io.spine.time.Time.MILLIS_PER_SECOND;
-import static io.spine.time.Time.NANOS_PER_MICROSECOND;
-import static io.spine.time.Time.NANOS_PER_SECOND;
-import static io.spine.time.Time.SECONDS_PER_HOUR;
-import static io.spine.time.Time.getCurrentTime;
-import static io.spine.time.Time.resetProvider;
-import static io.spine.time.Time.setProvider;
-import static io.spine.time.Time.systemTime;
+import static io.spine.time.EarthTime.HOURS_PER_DAY;
+import static io.spine.time.EarthTime.SECONDS_PER_HOUR;
 import static io.spine.time.Timestamps2.compare;
 import static io.spine.time.Timestamps2.isBetween;
 import static io.spine.time.Timestamps2.isLaterThan;
